requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-channels==4.0.0
channels-redis==4.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
celery==5.3.4
redis==5.0.1
Pillow==10.0.1
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
stripe==7.8.0
openai==1.3.5
websockets==12.0
channels-redis==4.1.0
daphne==4.0.0